Refactor navigation and social data structure for improved organization and consistency
