+++ data/socials.ts
